wandb_mnist/requirements.txt
torch
torchvision
wandb
numpy
pandas
matplotlib
scipy
tqdm
pyyaml
Fire
scikit-learn